+++ src/main/java/org/organicdesign/fp/collections/UnmodSortedCollection.java
@@ -1,6 +1,9 @@
 package org.organicdesign.fp.collections;
 
+import java.util.Set;
+
 public interface UnmodSortedCollection<E> extends UnmodCollection<E>, UnmodSortedIterable<E> {
+
     /** An unmodifiable ordered iterator {@inheritDoc} */
     @Override
     UnmodSortedIterator<E> iterator();