+++ src/test/java/org/organicdesign/fp/FunctionUtilsTest.java
@@ -29,7 +29,6 @@ import java.util.SortedMap;
 import java.util.TreeMap;
 import java.util.TreeSet;
 
-import org.junit.Assert;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.junit.runners.JUnit4;
@@ -76,43 +75,43 @@ public class FunctionUtilsTest {
         }
     }
 
-    @Test
-    public void testToString() {
-        List<Integer> is = new ArrayList<>();
-        is.add(1);
-        is.add(2);
-        is.add(3);
-        is.add(4);
-        is.add(5);
-        assertEquals("Array<Integer>(1,2,3,4,5)", FunctionUtils.arrayToString(is.toArray()));
+//    @Test
+//    public void testToString() {
+//        List<Integer> is = new ArrayList<>();
+//        is.add(1);
+//        is.add(2);
+//        is.add(3);
+//        is.add(4);
+//        is.add(5);
+//        assertEquals("Array<Integer>(1,2,3,4,5)", FunctionUtils.arrayToString(is.toArray()));
+//
+//        is.add(6);
+//        assertEquals("Array<Integer>(1,2,3,4,5,...)", FunctionUtils.arrayToString(is.toArray()));
+//
+//        Map<String,Integer> m = new TreeMap<>();
+//        m.put("Hello", 99);
+//        m.put("World", -237);
+//        assertEquals("TreeMap(Entry(Hello,99),Entry(World,-237))", FunctionUtils.mapToString(m));
+//
+//        m.put("x", 3);
+//        m.put("y", 2);
+//        m.put("z", 1);
+//        m.put("zz", 0);
+//
+//        assertEquals("TreeMap(Entry(Hello,99),Entry(World,-237),Entry(x,3),Entry(y,2),Entry(z,1),...)",
+//                     FunctionUtils.mapToString(m));
+//
+//        assertEquals("Array()", FunctionUtils.arrayToString(new Integer[0]));
+//        assertEquals("Array(null)", FunctionUtils.arrayToString(new Integer[] {null}));
+//
+//    }
 
-        is.add(6);
-        assertEquals("Array<Integer>(1,2,3,4,5,...)", FunctionUtils.arrayToString(is.toArray()));
-
-        Map<String,Integer> m = new TreeMap<>();
-        m.put("Hello", 99);
-        m.put("World", -237);
-        assertEquals("TreeMap(Entry(Hello,99),Entry(World,-237))", FunctionUtils.mapToString(m));
-
-        m.put("x", 3);
-        m.put("y", 2);
-        m.put("z", 1);
-        m.put("zz", 0);
-
-        assertEquals("TreeMap(Entry(Hello,99),Entry(World,-237),Entry(x,3),Entry(y,2),Entry(z,1),...)",
-                     FunctionUtils.mapToString(m));
-
-        assertEquals("Array()", FunctionUtils.arrayToString(new Integer[0]));
-        assertEquals("Array(null)", FunctionUtils.arrayToString(new Integer[] {null}));
-
-    }
-
-    @SuppressWarnings({"ConstantConditions","Unchecked"})
-    @Test public void testToStringNull() {
-        assertEquals("null", FunctionUtils.mapToString(null));
-        Integer[] zs = null;
-        assertEquals("null", FunctionUtils.arrayToString(zs));
-    }
+//    @SuppressWarnings({"ConstantConditions","Unchecked"})
+//    @Test public void testToStringNull() {
+//        assertEquals("null", FunctionUtils.mapToString(null));
+//        Integer[] zs = null;
+//        assertEquals("null", FunctionUtils.arrayToString(zs));
+//    }
 
 //    @SuppressWarnings("Convert2Lambda")
 //    public static final Predicate<Integer> r = new Predicate<Integer>() {
@@ -330,19 +329,19 @@ public class FunctionUtilsTest {
                                   "Sixteen","Seventeen","Eighteen","Nineteen","Twenty"};
 
     public static void mapHelper(Map<Integer,String> m, int max) {
-        Assert.assertEquals("Size check", max, m.size());
+        assertEquals("Size check", max, m.size());
         for (int i = 0; i < max; i++) {
-            Assert.assertEquals(ss[i], m.get(i + 1));
+            assertEquals(ss[i], m.get(i + 1));
         }
         assertNull(m.get(max + 1));
         assertNull(m.get(max + 999));
     }
 
     public static void mapHelperOdd(Map<Integer,String> m, int max) {
-        Assert.assertEquals("Size check", (max + 1) / 2, m.size());
+        assertEquals("Size check", (max + 1) / 2, m.size());
         for (int i = 0; i < max; i++) {
             if ( (i % 2) == 0 ) {
-                Assert.assertEquals(ss[i], m.get(i + 1));
+                assertEquals(ss[i], m.get(i + 1));
             } else {
                 assertNull(m.get(i + 1));
             }
@@ -352,12 +351,12 @@ public class FunctionUtilsTest {
     }
 
     public static void mapHelperEven(Map<Integer,String> m, int max) {
-        Assert.assertEquals("Size check", max / 2, m.size());
+        assertEquals("Size check", max / 2, m.size());
         for (int i = 0; i < max; i++) {
             if ( (i % 2) == 0 ) {
                 assertNull(m.get(i + 1));
             } else {
-                Assert.assertEquals(ss[i], m.get(i + 1));
+                assertEquals(ss[i], m.get(i + 1));
             }
         }
         assertNull(m.get(max + 1));
@@ -563,24 +562,24 @@ public class FunctionUtilsTest {
 //        assertEquals(UnmodSet.empty(), unSetSkipNull());
 //    }
 
-    @Test public void emptyIteratorTest() {
-        assertTrue(emptyUnmodIterable() == unmodIterable(null));
-        assertTrue(emptyUnmodIterable() == serializeDeserialize(emptyUnmodIterable()));
-        assertFalse(emptyUnmodIterable().iterator().hasNext());
-    }
-
-    @Test(expected = NoSuchElementException.class)
-    public void emptyIteratorTestEx() { emptyUnmodIterable().iterator().next(); }
-
-
-    @Test public void unmodIterableTest() {
-        ImList<Integer> oneTwoThree = vec(1,2,3);
-        assertTrue("An unmod iterable comes through unmodified",
-                   oneTwoThree == unmodIterable(oneTwoThree));
-
-        TestUtilities.iteratorTest(Arrays.asList(1, 2, 3).iterator(),
-                                   unmodIterable(Arrays.asList(1,2,3)).iterator());
-    }
+//    @Test public void emptyIteratorTest() {
+//        assertFalse(unmodIterable(null).iterator().hasNext());
+//        assertFalse(emptyUnmodIterable().iterator().hasNext());
+//        assertFalse(serializeDeserialize(emptyUnmodIterable()).iterator().hasNext());
+//    }
+//
+//    @Test(expected = NoSuchElementException.class)
+//    public void emptyIteratorTestEx() { emptyUnmodIterable().iterator().next(); }
+//
+//
+//    @Test public void unmodIterableTest() {
+//        ImList<Integer> oneTwoThree = vec(1,2,3);
+//        assertTrue("An unmod iterable comes through unmodified",
+//                   oneTwoThree == unmodIterable(oneTwoThree));
+//
+//        TestUtilities.iteratorTest(Arrays.asList(1, 2, 3).iterator(),
+//                                   unmodIterable(Arrays.asList(1,2,3)).iterator());
+//    }
 
     @Test public void testEmptyUnmodListIterator() {
         assertFalse(emptyUnmodListIterator().hasNext());
@@ -590,21 +589,20 @@ public class FunctionUtilsTest {
     }
 
     @Test (expected = NoSuchElementException.class)
-    public void testUnmodListIteratorNext() { EMPTY_UNMOD_LIST_ITERATOR.next(); }
+    public void testUnmodListIteratorNext() { emptyUnmodListIterator().next(); }
 
     @Test (expected = NoSuchElementException.class)
-    public void testUnmodListIteratorPrev() { EMPTY_UNMOD_LIST_ITERATOR.previous(); }
+    public void testUnmodListIteratorPrev() { emptyUnmodListIterator().previous(); }
 
     @Test public void unmodListIteratorTest() {
-        UnmodList emptyUnList = FunctionUtils.unmodList(null);
-        assertTrue(FunctionUtils.emptyUnmodList() == emptyUnList);
-        // TODO: Make this pass!
-        assertTrue(FunctionUtils.emptyUnmodList() == serializeDeserialize(emptyUnList));
-        TestUtilities.listIteratorTest(Collections.emptyList(), FunctionUtils.emptyUnmodList());
+        UnmodList emptyUnList = unmodList(null);
+        assertEquals(emptyUnmodList(), emptyUnList);
+        assertEquals(emptyUnmodList(), serializeDeserialize(emptyUnList));
+        TestUtilities.listIteratorTest(Collections.emptyList(), emptyUnmodList());
 
-        assertTrue(FunctionUtils.emptyUnmodList() == FunctionUtils.unmodList(Collections.emptyList()));
-        assertTrue(FunctionUtils.emptyUnmodList() ==
-                   serializeDeserialize(FunctionUtils.unmodList(Collections.emptyList())));
+        assertEquals(emptyUnmodList(), unmodList(Collections.emptyList()));
+        assertEquals(emptyUnmodList(),
+                   serializeDeserialize(unmodList(Collections.emptyList())));
 
         ImList<Integer> oneTwoThree = vec(1,2,3);
         assertTrue("An unmod List comes through unmodified",
@@ -614,8 +612,8 @@ public class FunctionUtilsTest {
     }
 
     @Test public void unListTest() {
-        equalsDistinctHashCode(FunctionUtils.unmodList(Arrays.asList(3, 4, 5)),
-                               FunctionUtils.unmodList(new ArrayList<>(Arrays.asList(3, 4, 5))),
+        equalsDistinctHashCode(unmodList(Arrays.asList(3, 4, 5)),
+                               unmodList(new ArrayList<>(Arrays.asList(3, 4, 5))),
                                new LinkedList<>(Arrays.asList(3, 4, 5)),
                                new ArrayList<>(Arrays.asList(4, 5, 3)));
 
@@ -655,38 +653,45 @@ public class FunctionUtilsTest {
         assertFalse(emptyUnmodSet().contains(null));
         assertEquals(0, emptyUnmodSet().size());
         assertTrue(emptyUnmodSet().isEmpty());
-        assertTrue(emptyUnmodIterator() == emptyUnmodSet().iterator());
+        assertFalse(emptyUnmodSet().iterator().hasNext());
     }
 
     @Test public void unSetTest() {
-        UnmodSet<Integer> s = FunctionUtils.unmodSet(new HashSet<>(Arrays.asList(5, 4, 3)));
+        UnmodSet<Integer> s = unmodSet(new HashSet<>(Arrays.asList(5, 4, 3)));
 
-        assertEquals(FunctionUtils.EMPTY_UNMOD_SET, FunctionUtils.unmodSet(null));
-        assertEquals(FunctionUtils.EMPTY_UNMOD_SET, FunctionUtils.unmodSet(Collections.emptySet()));
+        assertTrue(unmodSet(null).isEmpty());
+        assertFalse(unmodSet(null).iterator().hasNext());
+        assertEquals(emptyUnmodSet(), unmodSet(Collections.emptySet()));
 
         ImSet<Integer> imSet = set(1,2,3);
-        assertTrue(imSet == FunctionUtils.unmodSet(imSet));
+        assertTrue(imSet == unmodSet(imSet));
 
         assertTrue(s.contains(3));
         assertFalse(s.contains(-1));
         assertFalse(s.isEmpty());
-        assertTrue(FunctionUtils.unmodSet(Collections.emptySet()).isEmpty());
+        assertTrue(unmodSet(Collections.emptySet()).isEmpty());
 
         equalsDistinctHashCode(s,
-                               FunctionUtils.unmodSet(new HashSet<>(Arrays.asList(3, 4, 5))),
+                               unmodSet(new HashSet<>(Arrays.asList(3, 4, 5))),
                                new HashSet<>(Arrays.asList(4, 3, 5)),
-                               FunctionUtils.unmodSet(new HashSet<>(Arrays.asList(4, 5, 6)))
+                               unmodSet(new HashSet<>(Arrays.asList(4, 5, 6)))
         );
     }
 
     @Test public void emptyUnmodSortedSetTest() {
-        assertFalse(emptyUnmodSortedSet().contains(null));
-        Assert.assertEquals(0, emptyUnmodSortedSet().size());
+// Not valid arguments.
+//        assertFalse(emptyUnmodSortedSet().contains(null));
+        assertFalse(emptyUnmodSortedSet().contains(""));
+        assertFalse(emptyUnmodSortedSet().contains(0));
+        assertFalse(emptyUnmodSortedSet().contains(1));
+        assertFalse(emptyUnmodSortedSet().contains((Comparable) o -> 0));
+        assertEquals(0, emptyUnmodSortedSet().size());
         assertTrue(emptyUnmodSortedSet().isEmpty());
-        assertTrue(EMPTY_UNMOD_SORTED_ITERATOR == emptyUnmodSortedSet().iterator());
+        assertFalse(emptyUnmodSortedSet().iterator().hasNext());
         assertNull(emptyUnmodSortedSet().comparator());
-        assertTrue(EMPTY_UNMOD_SORTED_SET == emptyUnmodSortedSet().subSet(null, null));
-        assertTrue(EMPTY_UNMOD_SORTED_SET == emptyUnmodSortedSet().tailSet(null));
+// Not valid arguments.
+//        assertTrue(EMPTY_UNMOD_SORTED_SET == emptyUnmodSortedSet().subSet(null, null));
+//        assertTrue(EMPTY_UNMOD_SORTED_SET == emptyUnmodSortedSet().tailSet(null));
     }
 
     @Test (expected = NoSuchElementException.class)
@@ -697,14 +702,14 @@ public class FunctionUtilsTest {
 
 
     @Test public void unSetSorted() {
-        assertEquals(FunctionUtils.EMPTY_UNMOD_SORTED_SET, FunctionUtils.unmodSortedSet(null));
-        assertEquals(FunctionUtils.EMPTY_UNMOD_SORTED_SET,
-                     FunctionUtils.unmodSortedSet(Collections.emptySortedSet()));
+        assertEquals(emptyUnmodSortedSet(), unmodSortedSet(null));
+        assertEquals(emptyUnmodSortedSet(),
+                     unmodSortedSet(Collections.emptySortedSet()));
 
         ImSortedSet<Integer> imSet = sortedSet((a, b) -> a - b, vec(1, 2, 3));
-        assertTrue(imSet == FunctionUtils.unmodSortedSet(imSet));
+        assertTrue(imSet == unmodSortedSet(imSet));
 
-        UnmodSortedSet<Integer> ts = FunctionUtils.unmodSortedSet(new TreeSet<>(Arrays.asList(5, 4, 3)));
+        UnmodSortedSet<Integer> ts = unmodSortedSet(new TreeSet<>(Arrays.asList(5, 4, 3)));
         assertNull(ts.comparator());
         // headSet is exclusive.
         assertTrue(ts.headSet(4).contains(3));
@@ -716,8 +721,8 @@ public class FunctionUtilsTest {
         assertTrue(ts.tailSet(4).contains(4));
         assertFalse(ts.tailSet(4).contains(3));
 
-        Assert.assertEquals(Integer.valueOf(3), ts.first());
-        Assert.assertEquals(Integer.valueOf(5), ts.last());
+        assertEquals(Integer.valueOf(3), ts.first());
+        assertEquals(Integer.valueOf(5), ts.last());
 
         assertFalse(ts.contains(2));
         assertTrue(ts.contains(3));
@@ -732,25 +737,26 @@ public class FunctionUtilsTest {
 
         assertFalse(ts.isEmpty());
 
-        Assert.assertEquals(ts.hashCode(), FunctionUtils.unmodSortedSet(new TreeSet<>(Arrays.asList(5, 4, 3))).hashCode());
-        Assert.assertEquals(ts, FunctionUtils.unmodSortedSet(new TreeSet<>(Arrays.asList(5, 4, 3))));
+        assertEquals(ts.hashCode(), unmodSortedSet(new TreeSet<>(Arrays.asList(5, 4, 3))).hashCode());
+        assertEquals(ts, unmodSortedSet(new TreeSet<>(Arrays.asList(5, 4, 3))));
 
-        equalsDistinctHashCode(FunctionUtils.unmodSortedSet(new TreeSet<>(Arrays.asList(5, 4, 3))),
-                               FunctionUtils.unmodSortedSet(new TreeSet<>(Arrays.asList(3, 4, 5))),
+        equalsDistinctHashCode(unmodSortedSet(new TreeSet<>(Arrays.asList(5, 4, 3))),
+                               unmodSortedSet(new TreeSet<>(Arrays.asList(3, 4, 5))),
                                new TreeSet<>(Arrays.asList(4, 3, 5)),
-                               FunctionUtils.unmodSortedSet(new TreeSet<>(Arrays.asList(4, 5, 6)))
+                               unmodSortedSet(new TreeSet<>(Arrays.asList(4, 5, 6)))
         );
     }
 
     @SuppressWarnings("deprecation")
     @Test public void emptyUnmodMapTest() {
         assertEquals(0, emptyUnmodMap().entrySet().size());
-        assertTrue(EMPTY_UNMOD_SET == emptyUnmodMap().keySet());
-        assertTrue(EMPTY_UNMOD_COLLECTION == emptyUnmodMap().values());
+        assertEquals(emptyUnmodSet(), emptyUnmodMap().keySet());
+        // Can't test equality on collections.
+        assertTrue(emptyUnmodMap().values().isEmpty());
+        assertFalse(emptyUnmodMap().values().iterator().hasNext());
         assertEquals(0, emptyUnmodMap().size());
         assertTrue(emptyUnmodMap().isEmpty());
-        assertTrue(FunctionUtils.<UnmodMap.UnEntry<Object,Object>>emptyUnmodIterator() ==
-                   emptyUnmodMap().iterator());
+        assertFalse(emptyUnmodMap().iterator().hasNext());
         assertFalse(emptyUnmodMap().containsKey(null));
         assertFalse(emptyUnmodMap().containsValue(null));
         assertNull(emptyUnmodMap().get(null));
@@ -758,24 +764,24 @@ public class FunctionUtilsTest {
 
     @SuppressWarnings("deprecation")
     @Test public void unMapTest() {
-        assertEquals(FunctionUtils.EMPTY_UNMOD_MAP, FunctionUtils.unmodMap(null));
-        assertEquals(FunctionUtils.EMPTY_UNMOD_MAP,
-                     FunctionUtils.unmodMap(Collections.emptyMap()));
+        assertEquals(EMPTY_UNMOD_MAP, unmodMap(null));
+        assertEquals(EMPTY_UNMOD_MAP,
+                     unmodMap(Collections.emptyMap()));
 
         ImMap<Integer,String> imMap = map(tup(1, ordinal(1)),
                                           tup(2, ordinal(2)),
                                           tup(3, ordinal(3)));
 
-        assertTrue(imMap == FunctionUtils.unmodMap(imMap));
+        assertTrue(imMap == unmodMap(imMap));
 
         final UnmodMap<Integer,String> ts;
         Map<Integer,String> sm = new HashMap<>();
         sm.put(5, "five");
         sm.put(4, "four");
         sm.put(3, "three");
-        ts = FunctionUtils.unmodMap(sm);
+        ts = unmodMap(sm);
 
-        Assert.assertEquals(3, ts.size());
+        assertEquals(3, ts.size());
         assertFalse(ts.isEmpty());
 
         assertFalse(ts.containsKey(2));
@@ -804,7 +810,7 @@ public class FunctionUtilsTest {
             sm2.put(3, "three");
             sm2.put(4, "four");
             sm2.put(5, "five");
-            m2 = FunctionUtils.unmodMap(sm2);
+            m2 = unmodMap(sm2);
         }
 
         final UnmodMap<Integer,String> m3;
@@ -813,51 +819,50 @@ public class FunctionUtilsTest {
             sm3.put(4, "four");
             sm3.put(5, "five");
             sm3.put(6, "six");
-            m3 = FunctionUtils.unmodMap(sm3);
+            m3 = unmodMap(sm3);
         }
 
         equalsDistinctHashCode(ts, m2, sm, m3);
 
-        Assert.assertEquals(3, ts.entrySet().size());
+        assertEquals(3, ts.entrySet().size());
         assertFalse(ts.entrySet().isEmpty());
 
-        Assert.assertEquals(3, ts.keySet().size());
+        assertEquals(3, ts.keySet().size());
         assertFalse(ts.keySet().isEmpty());
 
-        Assert.assertEquals(3, ts.values().size());
+        assertEquals(3, ts.values().size());
         assertFalse(ts.values().isEmpty());
 
         equalsDistinctHashCode(ts.entrySet(), m2.entrySet(), sm.entrySet(), m3.entrySet());
         equalsDistinctHashCode(ts.keySet(), m2.keySet(), sm.keySet(), m3.keySet());
 
-        Assert.assertEquals(m3, m3);
+        assertEquals(m3, m3);
 
         // Wow.  HashMap.values() returns something that doesn't implement equals.
 //        assertEquals(m3.values(), m3.values());
-        Assert.assertEquals(new ArrayList<>(m3.values()), new ArrayList<>(m3.values()));
+        assertEquals(new ArrayList<>(m3.values()), new ArrayList<>(m3.values()));
 
         equalsDistinctHashCode(new ArrayList<>(ts.values()), new ArrayList<>(m2.values()),
                                new ArrayList<>(sm.values()), new ArrayList<>(m3.values()));
     }
 
     @Test public void unMapSorted() {
-        assertEquals(FunctionUtils.EMPTY_UNMOD_SORTED_MAP, FunctionUtils.unmodSortedMap(null));
-        assertEquals(FunctionUtils.EMPTY_UNMOD_SORTED_MAP,
-                     FunctionUtils.unmodSortedMap(Collections.emptySortedMap()));
+        assertEquals(emptyUnmodSortedMap(), unmodSortedMap(null));
+        assertEquals(emptyUnmodSortedMap(), unmodSortedMap(Collections.emptySortedMap()));
 
         ImSortedMap<Integer,String> imMap = sortedMap((a, b) -> a - b, vec(tup(1, ordinal(1)),
                                                                            tup(2, ordinal(2)),
                                                                            tup(3, ordinal(3))));
-        assertTrue(imMap == FunctionUtils.unmodSortedMap(imMap));
+        assertTrue(imMap == unmodSortedMap(imMap));
 
         final UnmodSortedMap<Integer,String> ts;
         SortedMap<Integer,String> sm = new TreeMap<>();
         sm.put(5, "five");
         sm.put(4, "four");
         sm.put(3, "three");
-        ts = FunctionUtils.unmodSortedMap(sm);
+        ts = unmodSortedMap(sm);
 
-        Assert.assertEquals(3, ts.size());
+        assertEquals(3, ts.size());
         assertFalse(ts.isEmpty());
 
         assertNull(ts.comparator());
@@ -879,8 +884,8 @@ public class FunctionUtilsTest {
         assertTrue(ts.tailMap(4).containsValue("four"));
         assertFalse(ts.tailMap(4).containsValue("three"));
 
-        Assert.assertEquals(Integer.valueOf(3), ts.firstKey());
-        Assert.assertEquals(Integer.valueOf(5), ts.lastKey());
+        assertEquals(Integer.valueOf(3), ts.firstKey());
+        assertEquals(Integer.valueOf(5), ts.lastKey());
 
         assertFalse(ts.containsKey(2));
         assertTrue(ts.containsKey(3));
@@ -907,7 +912,7 @@ public class FunctionUtilsTest {
             sm2.put(3, "three");
             sm2.put(4, "four");
             sm2.put(5, "five");
-            m2 = FunctionUtils.unmodSortedMap(sm2);
+            m2 = unmodSortedMap(sm2);
         }
 
         final UnmodSortedMap<Integer,String> m3;
@@ -916,18 +921,18 @@ public class FunctionUtilsTest {
             sm3.put(4, "four");
             sm3.put(5, "five");
             sm3.put(6, "six");
-            m3 = FunctionUtils.unmodSortedMap(sm3);
+            m3 = unmodSortedMap(sm3);
         }
 
         equalsDistinctHashCode(ts, m2, sm, m3);
 
-        Assert.assertEquals(3, ts.entrySet().size());
+        assertEquals(3, ts.entrySet().size());
         assertFalse(ts.entrySet().isEmpty());
 
-        Assert.assertEquals(3, ts.keySet().size());
+        assertEquals(3, ts.keySet().size());
         assertFalse(ts.keySet().isEmpty());
 
-        Assert.assertEquals(3, ts.values().size());
+        assertEquals(3, ts.values().size());
         assertFalse(ts.values().isEmpty());
 
         equalsDistinctHashCode(ts.entrySet(), m2.entrySet(), sm.entrySet(), m3.entrySet());
@@ -935,11 +940,11 @@ public class FunctionUtilsTest {
 //        equalsDistinctHashCode(serializeDeserialize(ts.entrySet()), m2.entrySet(), sm.entrySet(), m3.entrySet());
         equalsDistinctHashCode(ts.keySet(), m2.keySet(), sm.keySet(), m3.keySet());
 
-        Assert.assertEquals(m3, m3);
+        assertEquals(m3, m3);
 
         // Wow.  TreeMap.values() returns something that doesn't implement equals.
 //        assertEquals(m3.values(), m3.values());
-        Assert.assertEquals(new ArrayList<>(m3.values()), new ArrayList<>(m3.values()));
+        assertEquals(new ArrayList<>(m3.values()), new ArrayList<>(m3.values()));
 
         equalsDistinctHashCode(new ArrayList<>(ts.values()), new ArrayList<>(m2.values()),
                                new ArrayList<>(sm.values()), new ArrayList<>(m3.values()));
@@ -973,18 +978,21 @@ public class FunctionUtilsTest {
     public void testEmptyUnmodSortedIteratorNext() { emptyUnmodSortedIterator().next(); }
 
     @Test public void testEmptyUnmodSortedMap() {
-        Assert.assertEquals(emptyUnmodSortedSet(), emptyUnmodSortedMap().entrySet());
-        Assert.assertEquals(emptyUnmodSortedSet(), emptyUnmodSortedMap().keySet());
+        assertEquals(emptyUnmodSortedSet(), emptyUnmodSortedMap().entrySet());
+        assertEquals(emptyUnmodSortedSet(), emptyUnmodSortedMap().keySet());
         assertNull(emptyUnmodSortedMap().comparator());
-        assertEquals(emptyUnmodSortedMap(), emptyUnmodSortedMap().subMap(null, null));
-        assertEquals(emptyUnmodSortedMap(), emptyUnmodSortedMap().tailMap(null));
-        Assert.assertEquals(emptyUnmodList(), emptyUnmodSortedMap().values());
-        Assert.assertEquals(0, emptyUnmodSortedMap().size());
+// Not valid arguments.
+//        assertEquals(emptyUnmodSortedMap(), emptyUnmodSortedMap().subMap(null, null));
+//        assertEquals(emptyUnmodSortedMap(), emptyUnmodSortedMap().tailMap(null));
+        assertTrue(emptyUnmodSortedMap().values().isEmpty());
+        assertFalse(emptyUnmodSortedMap().values().iterator().hasNext());
+        assertEquals(0, emptyUnmodSortedMap().size());
         assertTrue(emptyUnmodSortedMap().isEmpty());
-        Assert.assertEquals(emptyUnmodSortedIterator(), emptyUnmodSortedMap().iterator());
-        assertFalse(emptyUnmodSortedMap().containsKey(null));
+        assertTrue(emptyUnmodSortedMap().isEmpty());
+        assertFalse(emptyUnmodSortedMap().iterator().hasNext());
+        assertFalse(emptyUnmodSortedMap().containsKey((Comparable) o -> 0));
         assertFalse(emptyUnmodSortedMap().containsValue(null));
-        assertNull(emptyUnmodSortedMap().get(null));
+        assertNull(emptyUnmodSortedMap().get((Comparable) o -> 0));
     }
 
     @Test (expected = NoSuchElementException.class)
@@ -994,21 +1002,20 @@ public class FunctionUtilsTest {
 
 
     @SuppressWarnings("deprecation")
-    @Test public void unCollection() {
-        assertEquals(FunctionUtils.EMPTY_UNMOD_COLLECTION, FunctionUtils.unmodCollection(null));
-        assertEquals(FunctionUtils.EMPTY_UNMOD_COLLECTION,
-                     FunctionUtils.unmodCollection(Collections.emptySet()));
+    @Test public void testUnCollection() {
+        assertTrue(unmodCollection(null).isEmpty());
+        assertFalse(unmodCollection(null).iterator().hasNext());
 
         ImSet<Integer> imSet = set(1, 2, 3);
-        assertTrue(imSet == FunctionUtils.unmodCollection(imSet));
+        assertTrue(imSet == unmodCollection(imSet));
 
         ArrayDeque<Integer> ad = new ArrayDeque<>(Arrays.asList(1, 2, 3));
         UnmodCollection<Integer> a = unmodCollection(new ArrayDeque<>(Arrays.asList(1, 2, 3)));
-        Assert.assertEquals(3, a.size());
+        assertEquals(3, a.size());
         assertTrue(a.contains(2));
         assertFalse(a.isEmpty());
 
-        Assert.assertEquals(3, a.size());
+        assertEquals(3, a.size());
         assertTrue(a.contains(2));
         assertFalse(a.isEmpty());
 
@@ -1020,6 +1027,8 @@ public class FunctionUtilsTest {
 
         equalsDistinctHashCode(new ArrayList<>(a), new ArrayList<>(ad), Arrays.asList(1, 2, 3),
                                Arrays.asList(3, 2, 1));
+
+        assertEquals("UnmodCollection(3,2)", unmodCollection(Arrays.asList(3,2)).toString());
     }
 
 
@@ -1042,7 +1051,7 @@ public class FunctionUtilsTest {
         assertFalse(emptyUnmodCollection().contains(null));
         assertEquals(0, emptyUnmodCollection().size());
         assertTrue(emptyUnmodCollection().isEmpty());
-        assertTrue(emptyUnmodIterator() == emptyUnmodCollection().iterator());
+        assertFalse(emptyUnmodCollection().iterator().hasNext());
     }
 
 }